+++ src/js/utils/WebColors.d.ts
@@ -1,0 +1,15 @@
+/**
+ * A set of 'beautiful' web colors (I know, beauty is in the eye of the beholder).
+ *
+ * I found this color chart with 11 colors and found them somewhat nice
+ *    https://www.pinterest.com/pin/229965124706497134/
+ *
+ * @requires Color
+ *
+ * @author   Ikaros Kappler
+ * @version  1.0.0
+ * @date     2018-11-10
+ * @modified 2020-10-23 Ported to Typescript.
+ **/
+import { Color } from "./datastructures/Color";
+export declare const WebColors: Color[];