+++ src/js/utils/datastructures/interfaces.js
@@ -1,0 +1,10 @@
+"use strict";
+/**
+ * Some often-used interfaces.
+ *
+ * @author  Ikaros Kappler
+ * @date    2020-10-23
+ * @version 1.0.0
+ **/
+Object.defineProperty(exports, "__esModule", { value: true });
+//# sourceMappingURL=interfaces.js.map