+++ src/ts/utils/WebColors.ts
@@ -1,4 +1,3 @@
-"use strict";
 /**
  * A set of 'beautiful' web colors (I know, beauty is in the eye of the beholder).
  *
@@ -12,12 +11,13 @@
  * @date     2018-11-10
  * @modified 2020-10-23 Ported to Typescript.
  **/
-Object.defineProperty(exports, "__esModule", { value: true });
-var Color_1 = require("./datastructures/Color");
+
+import { Color } from "./datastructures/Color";
+
 // This is the original order (like in the image).
 // But if you want to get clearly distinct 'random' colors in a web app, then
 // the order below is better mixed.
-/*
+/* 
 export const WebColors = [
     Color.makeRGB(255,67,55),  // Red
     Color.makeRGB(232,31,100), // Pink
@@ -32,17 +32,17 @@ export const WebColors = [
     Color.makeRGB(141,195,67)  // Light Green
 ];
 */
-exports.WebColors = [
-    Color_1.Color.makeRGB(1, 150, 137),
-    Color_1.Color.makeRGB(255, 67, 55),
-    Color_1.Color.makeRGB(6, 170, 245),
-    Color_1.Color.makeRGB(64, 81, 181),
-    Color_1.Color.makeRGB(232, 31, 100),
-    Color_1.Color.makeRGB(77, 175, 82),
-    Color_1.Color.makeRGB(156, 39, 175),
-    Color_1.Color.makeRGB(141, 195, 67),
-    Color_1.Color.makeRGB(103, 59, 184),
-    Color_1.Color.makeRGB(3, 189, 214),
-    Color_1.Color.makeRGB(35, 151, 245) // Blue
+
+export const WebColors = [
+    Color.makeRGB(1,150,137),  // Teal
+    Color.makeRGB(255,67,55),  // Red
+    Color.makeRGB(6,170,245),  // Light Blue
+    Color.makeRGB(64,81,181),  // Indigo
+    Color.makeRGB(232,31,100), // Pink
+    Color.makeRGB(77,175,82),  // Green
+    Color.makeRGB(156,39,175), // Purple
+    Color.makeRGB(141,195,67), // Light Green
+    Color.makeRGB(103,59,184), // Deep Purple
+    Color.makeRGB(3,189,214),  // Cyan
+    Color.makeRGB(35,151,245)  // Blue
 ];
-//# sourceMappingURL=WebColors.js.map